moving powerflow and overload graph creation to a common class independant of simulation class
